We now can compile as C++.

+++ src/act_obj.c
@@ -2038,7 +2038,7 @@ void do_auction (CHAR_DATA *ch, char *argument)
   else
     {
       act (AT_TELL, "Try again later - $p is being auctioned right now!",ch,auction->item,NULL,TO_CHAR);
-      WAIT_STATE( ch, 1.5 * PULSE_VIOLENCE );
+      WAIT_STATE( ch, (short)(1.5 * PULSE_VIOLENCE) );
       return;
     }
 }
